jcb-billing-extractor/requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
openpyxl==3.1.2
beautifulsoup4==4.12.3
